+++ src/timeapp.rs
@@ -262,18 +262,6 @@ mod tests {
     }
 
     #[test]
-    fn test_tsdec_verbose_cli_stdin() {
-        assert_cmd::Command::cargo_bin("rsbkb")
-            .expect("Could not run binary")
-            .args(&["tsdec", "-v"])
-            .write_stdin("1")
-            .assert()
-            .stdout("1970-01-01T00:00:01Z")
-            .stderr("Used format: Seconds since Epoch\n")
-            .success();
-    }
-
-    #[test]
     fn test_tsdec_decimal() {
         let ts = TsDecApplet {
             local: false,
@@ -312,45 +300,6 @@ mod tests {
 
     fn run_encode(app: &TsEncApplet, date: &str) -> String {
         String::from_utf8(app.process_test(date.as_bytes().to_vec())).unwrap()
-    }
-
-    #[test]
-    fn test_tsenc_cli_stdin() {
-        assert_cmd::Command::cargo_bin("rsbkb")
-            .expect("Could not run binary")
-            .args(&["tsenc"])
-            .write_stdin("1970-01-01T00:00:01Z")
-            .assert()
-            .stdout("1")
-            .success();
-        assert_cmd::Command::cargo_bin("rsbkb")
-            .expect("Could not run binary")
-            .args(&["tsenc", "-t", "filetime"])
-            .write_stdin("1601-01-01T00:00:01Z")
-            .assert()
-            .stdout("10000000")
-            .success();
-        assert_cmd::Command::cargo_bin("rsbkb")
-            .expect("Could not run binary")
-            .args(&["tsenc", "-i", "iso8601"])
-            .write_stdin("1970-01-01T00:00:01Z")
-            .assert()
-            .stdout("1")
-            .success();
-        assert_cmd::Command::cargo_bin("rsbkb")
-            .expect("Could not run binary")
-            .args(&["tsenc", "-i", "rfc2822"])
-            .write_stdin("Sat, 12 Jun 1993 13:25:19 GMT")
-            .assert()
-            .stdout("739891519")
-            .success();
-        assert_cmd::Command::cargo_bin("rsbkb")
-            .expect("Could not run binary")
-            .args(&["tsenc", "-i", "rfc3339"])
-            .write_stdin("1985-04-12T23:20:50.52Z")
-            .assert()
-            .stdout("482196050")
-            .success();
     }
 
     #[test]